Another waypoint commit
Currently the `init.el` file is the source of truth of how things
should be set up.

It has some issues right now but I'm currently more into the idea of
not doing a fully literate configuration after all. The disadvantages
of editing that much elisp not in an elisp buffer is a bit annoying.

Also continuing to try to minimalize bit by bit.